newitems.cgi - simple CGI script for linking to a list of new items in an Evergreen catalog.
It accepts a list of one or more Evergreen location IDs (passed via
the "loc") CGI parameter and an optional count.  It then fetches
the list of bib IDs of the $count newest items in those locations,
then generates an HTTP redirect to a TPAC search results page.

To use, place it on the Evergreen web server in a directory that's
configured to run CGI scripts.

An example of a URL using this script:

http://catalog.example.org/cgi-bin/newitems.cgi?count=30&loc=123&loc=124

Signed-off-by: Galen Charlton <gmc@esilibrary.com>